+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * #4546: Refactor GraphQL purgeAddress to permit multiple target addresses
 * #4549: Implement GraphQL closeConnections for brokered and standard
 * #4555: Enforce presence of routeServicePort in route endpoints definitions 
+* #4574: Set globalMaxSize to 1/4 of broker JVM heap
 
 ## 0.31.2
 * #4098: Added alert for high broker address memory usage